fix: fix ImageCache return type and simplify IdlixProvider cache usage
🔧 FIXED FINAL ERRORS:

1. ImageCache.kt - Fixed return type mismatch:
   - Added .toLong() to estimateWebPSize() result
   - Applied to all 6 providers

2. IdlixProvider.kt - Simplified cache usage (like Anichin):
   - Removed: CacheFingerprint, CacheValidationResult imports
   - Removed: fingerprints map
   - Simplified: CacheManager() without constructor params
   - Simplified: getMainPage() - no fingerprint checking
   - Using simple cache get/put API

✅ FILES MODIFIED:
- All ImageCache.kt files (6 providers)
- IdlixProvider/IdlixProvider.kt

🎯 EXPECTED: Build should now succeed!

+++ Donghuastream/src/main/kotlin/com/Donghuastream/ImageCache.kt
@@ -87,7 +87,7 @@ class ImageCache(
             quality >= 90 -> 0.35
             quality >= 80 -> 0.30
             else -> 0.25
-        })
+        }).toLong()
     }
 
     private suspend fun getCurrentTotalSize(): Long {
